move extension parsing into staticctapi
